socks: parameterise connection filter
Pass the config values that a SOCKS connection filter needs
when creating the filter: remote host/port and credentials.

Move the logic of finding out where a SOCKS filter should
tunnel into `connect.c` where the filter is created.

This keeps the interpretation of the various connection bits
in less places in the source and could allow to chain SOCKS
filters in future scenarios.