sftp/scp: try to break busy loop on receive when EAGAIN had been signalled
SFTP/SCP always did RUN_EXPIRE_NOW, causing a busy loop in state PERFORMING
because it was unsure of all received bytes had been processed by the protocol.

Regard EAGAIN on receive as a sign that all bytes were processed as far
as possible and that monitoring the socket alone is the safe thing to do.